backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-genai
python-dotenv==1.0.1
